feat: Handle permissions for utility costs

+++ resources/views/pages/utility-costs/⚡index/index.php
@@ -69,6 +69,8 @@ new class extends Component {
 
     public function save(): void
     {
+        $this->authorize('create', UtilityCost::class);
+
         $this->form->validate();
 
         $this->form->save();
@@ -80,6 +82,8 @@ new class extends Component {
 
     public function update(): void
     {
+        $this->authorize('update', UtilityCost::class);
+
         $this->form->validate();
 
         $this->form->update();
@@ -91,6 +95,8 @@ new class extends Component {
 
     public function deleteUtilityCost(int $id): void
     {
+        $this->authorize('delete', UtilityCost::class);
+
         $utilityCost = UtilityCost::findOrFail($id);
 
         $utilityCost->delete();
